+++ app/src/main/java/com/sw/wordgarden/data/datasource/remote/Retrofit/RetrofitClient.kt
@@ -1,4 +1,4 @@
-package com.sw.wordgarden.data.datasource.remote
+package com.sw.wordgarden.data.datasource.remote.Retrofit
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
